# docs/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: DataX Enterprise Studio API
  version: 1.2.0
  license:
    name: License pending repository-owner decision
    url: https://github.com/xiaoli2hust/datax/blob/main/README.md
  description: |
    V1 单组织、多项目、安全的一次性离线全量复制控制 API。
    仅支持 MySQL 8/PostgreSQL 15、人工触发、源端静默确认、预存在且实测为空的目标表和 insert-only-once。
    不提供调度、DAG、告警、AI、CDC、自由 SQL、Transformer 或插件安装接口。
    交付形态为 Windows 11 x64 本地工作站；业务 API 仅经 web 反向代理绑定本机
    127.0.0.1:17860，api、worker 和 postgres 不暴露宿主端口。
servers:
  - url: http://127.0.0.1:17860/api/v1
    description: Windows 本地工作站唯一支持入口；不得替换为 0.0.0.0、LAN IP 或公网地址
...
security:
  - bearerAuth: []
paths:
...
  /projects/{project_id}/members:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      operationId: listProjectMembers
      summary: 列出项目成员及角色
      description: 需要 Admin。
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 成员页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/members/{user_id}/roles:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [Projects]
      operationId: replaceProjectMemberRoles
      summary: 原子替换用户的项目角色
      description: 需要 Admin；角色可多选，权限取并集。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleReplaceRequest'
            example:
              roles: [DEVELOPER, VIEWER]
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /projects/{project_id}/transfer-policies:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [SecurityPolicies]
      operationId: listTransferPolicies
      summary: 分页列出项目源到目标传输策略
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 传输策略页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferPolicyPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [SecurityPolicies]
      operationId: createTransferPolicy
      summary: Admin 创建源修订到目标修订的 DRAFT 传输策略
      description: |
        请求必须分别给出 source/target 的 catalog/database、schema、table 和显式允许列。
        selection_mode=ALL_COLUMNS 也必须先由元数据解析为显式 allowed_columns 后提交，不接受
        *、正则或“运行时全部列”。服务端校验两侧表属于所选 DatasourceRevision，解析并固定
        PhysicalEndpointIdentity，按引擎标识符规则规范化/排序/去重后生成不可变 scope_json，
        再计算 scope_hash；客户端不能自报 scope_hash。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferPolicyCreate'
            example:
              source_datasource_revision_id: 11111111-1111-4111-8111-111111111112
              target_datasource_revision_id: 22222222-2222-4222-8222-222222222223
              requested_scope:
                schema_version: '1.0'
                source:
                  catalog: sales
                  schema: ''
                  table: orders
                  selection_mode: ALL_COLUMNS
                  allowed_columns: [amount, id]
                target:
                  catalog: warehouse
                  schema: public
                  table: orders_copy
                  selection_mode: SELECTED_COLUMNS
                  allowed_columns: [amount, id]
              classification: STANDARD
      responses:
        '201':
          description: 已创建 DRAFT
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferPolicy'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /transfer-policies/{transfer_policy_id}:
    parameters:
      - name: transfer_policy_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags: [SecurityPolicies]
      operationId: getTransferPolicy
      summary: 获取传输策略及脱敏审批摘要
      responses:
        '200':
          description: 传输策略
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferPolicy'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [SecurityPolicies]
      operationId: updateTransferPolicy
      summary: Admin 修订 DRAFT 或撤销策略
      description: 修改 source、target、requested_scope 或 classification 会把策略退回 DRAFT、清空旧审批并重算规范化 scope_json/scope_hash；ACTIVE 只允许变为 REVOKED。
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferPolicyPatch'
      responses:
        '200':
          description: 已更新
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferPolicy'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /transfer-policies/{transfer_policy_id}/submit:
    parameters:
      - name: transfer_policy_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    post:
      tags: [SecurityPolicies]
      operationId: submitTransferPolicy
      summary: Admin 提交策略审批
      description: expected_scope_hash 必须等于当前规范化 scope_hash，防止提交旧页面看到的范围。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferPolicySubmit'
            example:
              expected_scope_hash: 5f70bf18a08660b49f80c16fb9c0aa31c5e158d1471c4f4637ae0f0bde8f26f5
      responses:
        '200':
          description: 状态进入 PENDING_APPROVAL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferPolicy'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /transfer-policies/{transfer_policy_id}/approvals:
    parameters:
      - name: transfer_policy_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    post:
      tags: [SecurityPolicies]
      operationId: decideTransferPolicy
      summary: 不同 Admin 审批或拒绝传输策略
      description: STANDARD 至少一名 Admin 批准；SENSITIVE 至少两名不同 Admin 批准。申请人不得批准自己的策略。审批请求必须固定当前 expected_scope_hash；scope/revision/classification 任一修改即使旧审批失效。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferPolicyDecision'
            example:
              decision: APPROVED
              expected_scope_hash: 5f70bf18a08660b49f80c16fb9c0aa31c5e158d1471c4f4637ae0f0bde8f26f5
              comment: 已核对源目标表与显式列范围
      responses:
        '200':
          description: 当前策略；达到门槛后原子进入 ACTIVE，任一拒绝立即进入 REJECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferPolicy'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: des_refresh
      description: V1 loopback HTTP 使用 HttpOnly、SameSite=Strict、Path=/api/v1/auth 且无 Domain；若未来改为受信 loopback HTTPS 必须增加 Secure。
  headers:
...
    ETag:
      description: 弱 ETag，值为 W/"row_version"
      schema:
        type: string
        pattern: '^W/"[1-9][0-9]*"$'
...
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: 服务端生成的不透明游标
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^W/"[1-9][0-9]*"$'
  responses:
...
    Forbidden:
      description: 无权限
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 资源不存在或对当前主体不可见
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 版本、幂等、状态或唯一约束冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: 请求或业务语义校验失败
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    NullableTimestamp:
      oneOf:
        - $ref: '#/components/schemas/Timestamp'
        - type: 'null'
...
    Sha256:
      type: string
      pattern: '^[a-f0-9]{64}$'
    Role:
      type: string
      enum: [ADMIN, DEVELOPER, OPERATOR, VIEWER]
    ProjectRole:
      type: string
      enum: [DEVELOPER, OPERATOR, VIEWER]
...
    FieldError:
      type: object
      additionalProperties: false
      required: [path, code, message]
      properties:
        path:
          type: string
          maxLength: 500
        code:
          type: string
          pattern: '^[A-Z0-9_]+$'
        message:
          type: string
          maxLength: 1000
    CapacityAdmissionDetails:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          enum: [QUEUE_LIMIT, BACKLOG_LIMIT, DISK_YELLOW, DISK_RED, LOG_BUDGET]
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - code
        - detail
        - instance
        - request_id
        - retryable
        - field_errors
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z0-9_]+$'
        detail:
          type: [string, 'null']
          maxLength: 2000
        instance:
          type: string
          maxLength: 500
        request_id:
          $ref: '#/components/schemas/Uuid'
        retryable:
          type: boolean
        field_errors:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/FieldError'
        details:
          oneOf:
            - $ref: '#/components/schemas/CapacityAdmissionDetails'
            - type: 'null'
...
    UserSummary:
      type: object
      additionalProperties: false
      required: [id, email, display_name, must_change_password]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        email:
          type: string
          format: email
          maxLength: 254
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        must_change_password:
          type: boolean
          description: true 时除 refresh/logout/me/change-password 外的业务接口统一返回 PASSWORD_CHANGE_REQUIRED。
...
    Member:
      type: object
      additionalProperties: false
      required: [organization_member_id, user, roles]
      properties:
        organization_member_id:
          $ref: '#/components/schemas/Uuid'
        user:
          $ref: '#/components/schemas/UserSummary'
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Role'
    MemberPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, has_more]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
    RoleReplaceRequest:
      type: object
      additionalProperties: false
      required: [roles]
      properties:
        roles:
          type: array
          minItems: 0
          maxItems: 3
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ProjectRole'
...
    TransferPolicyApproval:
      type: object
      additionalProperties: false
      required: [id, approved_by, decision, comment, decided_at]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        approved_by:
          $ref: '#/components/schemas/Uuid'
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        comment:
          type: [string, 'null']
          maxLength: 500
        decided_at:
          $ref: '#/components/schemas/Timestamp'
    TransferPolicyScopeSide:
      type: object
      additionalProperties: false
      required:
        - physical_endpoint_identity_id
        - catalog
        - schema
        - table
        - allowed_columns
      properties:
        physical_endpoint_identity_id:
          $ref: '#/components/schemas/Uuid'
        catalog:
          type: string
          minLength: 1
          maxLength: 128
          description: 引擎原生 catalog/database；MySQL 为 database，PostgreSQL 为连接固定的 database/catalog。
        schema:
          type: string
          maxLength: 128
          description: PostgreSQL schema；MySQL 使用规范定义的空哨兵。
        table:
          type: string
          minLength: 1
          maxLength: 128
        allowed_columns:
          type: array
          minItems: 1
          maxItems: 2048
          uniqueItems: true
          description: 按引擎真实标识符规则规范化、排序和去重的显式列；不接受 * 或正则。
          items:
            type: string
            minLength: 1
            maxLength: 128
    TransferPolicyScopeInputSide:
      type: object
      additionalProperties: false
      description: Admin/UI 提交的单侧精确范围；PhysicalEndpointIdentity 由服务端从所选 DatasourceRevision 解析并校验。
      required: [catalog, schema, table, selection_mode, allowed_columns]
      properties:
        catalog:
          type: string
          minLength: 1
          maxLength: 128
          description: 引擎原生 catalog/database；MySQL 为 database，PostgreSQL 为连接固定的 database/catalog。
        schema:
          type: string
          maxLength: 128
          description: PostgreSQL schema；MySQL 使用规范定义的空哨兵。
        table:
          type: string
          minLength: 1
          maxLength: 128
        selection_mode:
          type: string
          enum: [ALL_COLUMNS, SELECTED_COLUMNS]
        allowed_columns:
          type: array
          minItems: 1
          maxItems: 2048
          uniqueItems: true
          description: 两种 selection_mode 都必须提交经当前元数据展开、按引擎规则规范化并排序去重的显式列；不接受 *、正则或延迟解析。
          items:
            type: string
            minLength: 1
            maxLength: 128
    TransferPolicyScopeInput:
      type: object
      additionalProperties: false
      required: [schema_version, source, target]
      properties:
        schema_version:
          const: '1.0'
        source:
          $ref: '#/components/schemas/TransferPolicyScopeInputSide'
        target:
          $ref: '#/components/schemas/TransferPolicyScopeInputSide'
    TransferPolicyScope:
      type: object
      additionalProperties: false
      required: [schema_version, source, target]
      properties:
        schema_version:
          const: '1.0'
        source:
          $ref: '#/components/schemas/TransferPolicyScopeSide'
        target:
          $ref: '#/components/schemas/TransferPolicyScopeSide'
    TransferPolicy:
      type: object
      additionalProperties: false
      required:
        - id
        - project_id
        - source_datasource_revision_id
        - target_datasource_revision_id
        - source_physical_endpoint_identity_id
        - target_physical_endpoint_identity_id
        - scope_json
        - scope_hash
        - classification
        - status
        - requested_by
        - approvals
        - activated_at
        - row_version
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        project_id:
          $ref: '#/components/schemas/Uuid'
        source_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        target_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        source_physical_endpoint_identity_id:
          $ref: '#/components/schemas/Uuid'
        target_physical_endpoint_identity_id:
          $ref: '#/components/schemas/Uuid'
        scope_json:
          $ref: '#/components/schemas/TransferPolicyScope'
        scope_hash:
          $ref: '#/components/schemas/Sha256'
          description: SHA-256(UTF8("DXTRANSFERPOLICYv1\n") || RFC8785(scope_json))。
        classification:
          type: string
          enum: [STANDARD, SENSITIVE]
        status:
          type: string
          enum: [DRAFT, PENDING_APPROVAL, ACTIVE, REJECTED, REVOKED]
        requested_by:
          $ref: '#/components/schemas/Uuid'
        approvals:
          type: array
          maxItems: 16
          items:
            $ref: '#/components/schemas/TransferPolicyApproval'
        activated_at:
          $ref: '#/components/schemas/NullableTimestamp'
        row_version:
          type: integer
          minimum: 1
    TransferPolicyCreate:
      type: object
      additionalProperties: false
      required:
        - source_datasource_revision_id
        - target_datasource_revision_id
        - requested_scope
        - classification
      properties:
        source_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        target_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        requested_scope:
          $ref: '#/components/schemas/TransferPolicyScopeInput'
          description: 服务端校验并规范化为响应中的 scope_json/scope_hash；客户端不得自报 hash。
        classification:
          type: string
          enum: [STANDARD, SENSITIVE]
    TransferPolicyPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        source_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        target_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        requested_scope:
          $ref: '#/components/schemas/TransferPolicyScopeInput'
        classification:
          type: string
          enum: [STANDARD, SENSITIVE]
        status:
          const: REVOKED
    TransferPolicySubmit:
      type: object
      additionalProperties: false
      required: [expected_scope_hash]
      properties:
        expected_scope_hash:
          $ref: '#/components/schemas/Sha256'
    TransferPolicyDecision:
      type: object
      additionalProperties: false
      required: [decision, expected_scope_hash]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        expected_scope_hash:
          $ref: '#/components/schemas/Sha256'
          description: 必须等于当前策略 scope_hash；任何范围修订都会使旧审批请求冲突。
        comment:
          type: [string, 'null']
          maxLength: 500
    TransferPolicyPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, has_more]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransferPolicy'
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
...